requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
kagglehub>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
torch>=2.0.0
torchvision>=0.15.0
